refactor: extract AppBootstrapper and ClassRegistrar from App.php
- Extract AppBootstrapper: path constants, encoding, autoloader init,
  stdlib checks, I/O streams, version info
- Extract ClassRegistrar: auto-discovery, reflection-based instantiation,
  namespace extraction, type checking
- App.php reduced from 966 to 622 lines
- App::initiate() delegates to AppBootstrapper::boot()
- App::autoRegister() delegates to ClassRegistrar::register()
- App::call() made public (needed by AppBootstrapper)
- App::getClassLoader() uses ClassLoader::get() directly
- All existing tests pass unchanged

Closes #350
